\name{alignStart}
\alias{alignStart}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function to align start of a window
}
\description{
Aligns start time to nearest minute interval.}
\usage{
alignStart(winSize, start)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{winSize}{
  Window size in seconds.
  }
  \item{start}{
  The start time as a POSIXlt DateTime object.
  }
}
\seealso{
\code{\link{DateTimeClasses}}
}
\author{
Katherine Ellis
}

%% ~Make other sections like Warning with \section{Warning }{....} ~


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
