\name{clearFiles}
\alias{clearFiles}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Clear files
}
\description{
Deletes files in a directory
}
\usage{
clearFiles(dir)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dir}{
  Path to directory to clear.
  }
}

\author{
Katherine Ellis
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
