\name{computeEmissionProbs}
\alias{computeEmissionProbs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compute emission probabilities}
\description{
Function to compute the emission probablilities of an HMM, based on the first-level random forest classifier.
}
\usage{
computeEmissionProbs(rf)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{rf}{
  A random forest object.
  }
}
\author{
Katherine Ellis
}
\seealso{
\code{\link{trainModel}}
}
%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
