\name{computePriorProbs}
\alias{computePriorProbs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compute prior probabilities
}
\description{
Compute the prior probabilities of an HMM from a state sequence.
}
\usage{
computePriorProbs(stateSeq)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{stateSeq}{
  Vector of states.
  }
}
\value{
Vector of probabilities for each state.
}
\author{
Katherine Ellis
}
%% ~Make other sections like Warning with \section{Warning }{....} ~
\seealso{
\code{\link{trainModel}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
