\name{hmm}
\alias{hmm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Hidden Markov model
}
\description{
Used for loading HMM from a TLBC model.
}
\author{
Katherine Ellis
}

%% ~Make other sections like Warning with \section{Warning }{....} ~


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
