\name{isFeatureDirectory}
\alias{isFeatureDirectory}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Is feature directory?
}
\description{
Function to check if the directory is a feature directory or a raw data directory (by checking for existence of sub-directories).
}
\usage{
isFeatureDirectory(dir)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dir}{
   Path to a directory.
  }
}
\value{
\emph{TRUE}, if the directory is a feature directory. \emph{FALSE} otherwise.
}
\author{
Katherine Ellis
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
