\name{isInstanceFormat}
\alias{isInstanceFormat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Is instance format?
}
\description{
Function to check if the directory contains instance-level annotations or bout-level annotations (by checking for existence of sub-directories).
}
\usage{
isInstanceFormat(annotations)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{annotations}{
   Path to an annotations directory.
  }
}
\value{
\emph{TRUE}, if the directory contains instance-level annotations. \emph{FALSE} otherwise.
}
\author{
Katherine Ellis
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
