\name{rf}
\alias{rf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Random Forest
}
\description{
Used for loading RF from a TLBC model.
}
\author{
Katherine Ellis
}

%% ~Make other sections like Warning with \section{Warning }{....} ~


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
