\name{senseCamLabels}
\alias{senseCamLabels}
\alias{senseCamLabelsFile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
SenseCam Labels
}
\description{
Function to convert bout-level annotations from senseCam labels to 6 categories (can be either single file or directory).
}
\usage{
senseCamLabels(input, output)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{input}{
  Path to file or directory containing bout-level annotations.
  }
  \item{output}{
  Path to output directory.
  }
}
\author{
Katherine Ellis
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
