\name{winSize}
\alias{winSize}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Window Size
}
\description{
Used for loading window size from a TLBC model.
}
\author{
Katherine Ellis
}

%% ~Make other sections like Warning with \section{Warning }{....} ~


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
